requirements.txt
torch==1.13.1 -f https://download.pytorch.org/whl/cpu/torch_stable.html
transformers==4.26.1
timm==0.6.12
pillow==9.4.0